Review round 3

+++ extensions/parquet/table/src/main/java/io/deephaven/parquet/table/transfer/BooleanTransfer.java
@@ -19,7 +19,7 @@ final class BooleanTransfer extends PrimitiveTransfer<WritableByteChunk<Values>,
                                   int targetPageSize) {
         final int NUM_BIT_PACKED_BOOLEANS_PER_BYTE = 8;
         final int maxValuesPerPage = Math.toIntExact(Math.min(tableRowSet.size(),
-                (long) targetPageSize * (long) NUM_BIT_PACKED_BOOLEANS_PER_BYTE));
+                (long) targetPageSize * NUM_BIT_PACKED_BOOLEANS_PER_BYTE));
         final byte[] backingArray = new byte[maxValuesPerPage];
         return new BooleanTransfer(
                 columnSource,
